+++ data-browser/src/App.tsx
@@ -11,6 +11,7 @@ import { MetaSetter } from './components/MetaSetter';
 import { Toaster } from './components/Toaster';
 import { getSnowpackEnv, isDev } from './config';
 import { handleWarning, initBugsnag } from './helpers/handlers';
+import { loadDefaultStore } from './helpers/loadDefaultStore';
 import HotKeysWrapper from './components/HotKeyWrapper';
 import { AppSettingsContextProvider } from './helpers/AppSettings';
 import ErrorPage from './views/ErrorPage';
@@ -21,6 +22,8 @@ const store = new Store();
 store.setBaseUrl(window.location.origin);
 /** Setup bugsnag for error handling */
 const ErrorBoundary = initBugsnag();
+/** Load the Atomic Data default properties and classes to speed things up */
+loadDefaultStore(store);
 
 /** Entrypoint of the application. This is where providers go. */
 function App(): JSX.Element {