+++ data-browser/src/components/PropVal.tsx
@@ -5,19 +5,27 @@ import styled from 'styled-components';
 import AtomicLink from './Link';
 import { ValueForm } from './forms/ValueForm';
 import { ErrorLook } from './ResourceInline';
+import ValueComp from './ValueComp';
 
 type Props = {
   propertyURL: string;
   resource: Resource;
+  editable: boolean;
+  // If set to true, will render the properties in a left column, and the Values in the right one, but only on large screens.
+  columns?: boolean;
 };
 
-export const PropValRow = styled.div`
-  display: flex;
-  flex-direction: row;
+interface PropValRowProps {
+  columns?: boolean;
+}
+
+export const PropValRow = styled.div<PropValRowProps>`
   word-break: break-word;
-  @media screen and (max-width: 500px) {
-    margin-bottom: 0.5rem;
-    flex-direction: column;
+  margin-bottom: ${p => (p.columns ? 0 : '0.5rem')};
+
+  @media screen and (min-width: 500px) {
+    flex-direction: ${p => (p.columns ? 'row' : 'column')};
+    display: ${p => (p.columns ? 'flex' : 'block')};
   }
 `;
 
@@ -31,7 +39,12 @@ export const PropertyLabel = styled.span`
  * A single Property / Value renderer that shows a label on the left, and the
  * value on the right. The value is editable.
  */
-function PropVal({ propertyURL, resource }: Props): JSX.Element {
+function PropVal({
+  propertyURL,
+  resource,
+  editable,
+  columns,
+}: Props): JSX.Element {
   const property = useProperty(propertyURL);
 
   if (property == null) {
@@ -41,7 +54,7 @@ function PropVal({ propertyURL, resource }: Props): JSX.Element {
   const truncated = truncateUrl(propertyURL, 10, true);
 
   return (
-    <PropValRow>
+    <PropValRow columns={columns}>
       <AtomicLink subject={propertyURL}>
         <PropertyLabel title={property.description}>
           {property.error ? (
@@ -52,7 +65,14 @@ function PropVal({ propertyURL, resource }: Props): JSX.Element {
           :
         </PropertyLabel>
       </AtomicLink>
-      <ValueForm resource={resource} propertyURL={propertyURL} />
+      {editable ? (
+        <ValueForm resource={resource} propertyURL={propertyURL} />
+      ) : (
+        <ValueComp
+          datatype={property.datatype}
+          value={resource.get(propertyURL)}
+        />
+      )}
     </PropValRow>
   );
 }