+++ data-browser/src/components/ResourcePage.tsx
@@ -79,7 +79,12 @@ function ResourcePage({ subject }: Props): JSX.Element {
       <h1>{title}</h1>
       <ClassDetail resource={resource} />
       <ValueForm resource={resource} propertyURL={properties.description} />
-      <AllProps resource={resource} except={defaulHiddenProps} />
+      <AllProps
+        resource={resource}
+        except={defaulHiddenProps}
+        editable
+        columns
+      />
       {/* Perhaps this should be an extendible runtime thing, where Classes have potential Actions. */}
       {klass == urls.classes.class && <NewInstanceButton klass={subject} />}
     </ContainerNarrow>