+++ data-browser/src/components/datatypes/NestedResource.tsx
@@ -1,13 +1,23 @@
 import React from 'react';
-import { NestedResource } from '@tomic/lib';
+import { Resource } from '@tomic/lib';
+import AllProps from '../AllProps';
+import styled from 'styled-components';
 
 type Props = {
-  resource: NestedResource;
+  resource: Resource;
 };
 
 /** Renders a Date value */
 function Nestedresource({ resource }: Props): JSX.Element {
-  return <div title='Nested Resource'>{JSON.stringify(resource)}</div>;
+  return (
+    <NestedWrapper>
+      <AllProps resource={resource} />
+    </NestedWrapper>
+  );
 }
+
+const NestedWrapper = styled.div`
+  margin-left: ${p => p.theme.margin}rem;
+`;
 
 export default Nestedresource;