+++ lib/src/parse.ts
@@ -1,17 +1,22 @@
 import { Resource, ResourceStatus } from './resource';
-import { Value } from './value';
+import { JSONObject, Value } from './value';
 
-/** Parses an JSON-AD string containing a resoure, adds it to the input Resource */
-export function parseJsonADResource(string: string, resource: Resource): void {
+/** Parses an JSON-AD object containing a resoure, adds it to the input Resource */
+export function parseJsonADResource(
+  jsonObject: JSONObject,
+  resource: Resource,
+): void {
   try {
-    const jsonObject = JSON.parse(string);
     for (const key in jsonObject) {
       if (key == '@id') {
-        const subject: string = jsonObject['@id'];
+        const subject = jsonObject['@id'];
         if (typeof subject !== 'string') {
           throw new Error("'@id' field must be a string");
         }
-        if (subject !== resource.getSubject()) {
+        if (
+          resource.getSubject() !== 'undefined' &&
+          subject !== resource.getSubject()
+        ) {
           throw new Error(
             `Resource has wrong subject in @id. Received subject was ${subject}, expected ${resource.getSubject()}.`,
           );
@@ -20,12 +25,12 @@ export function parseJsonADResource(string: string, resource: Resource): void {
         continue;
       }
       try {
-        const val = new Value(jsonObject[key]);
+        const path = resource.getSubject() + ' ' + key;
+        const val = new Value(jsonObject[key], path);
         resource.setUnsafe(key, val);
       } catch (e) {
         throw new Error(
-          `Failed creating value for key ${key} in resource ${resource.getSubject()}. ${
-            e.message
+          `Failed creating value for key ${key} in resource ${resource.getSubject()}. ${e.message
           }`,
         );
       }
@@ -37,4 +42,21 @@ export function parseJsonADResource(string: string, resource: Resource): void {
     throw e;
   }
   return;
+}
+
+/** Parsees a JSON-AD array string, returns array of Resources */
+export function parseJsonADArray(jsonArray: any[]): Resource[] {
+  const resources: Resource[] = [];
+  try {
+    // const jsonArray = JSON.parse(string);
+    for (const jsonObject of jsonArray) {
+      const resource = new Resource();
+      parseJsonADResource(jsonObject, resource);
+      resources.push(resource);
+    }
+  } catch (e) {
+    e.message = 'Failed parsing JSON ' + e.message;
+    throw e;
+  }
+  return resources;
 }